Ripping out old junk...
Moved old initialization code from app_UCM and OpenADRClient to Main.
